Separate MyExpander from guiutils.h
