+++ rtgui/tools/colorappearance.cc
@@ -26,6 +26,7 @@
 #include "options.h"
 #include "rtimage.h"
 #include "eventmapper.h"
+#include "widgets/basic/myexpander.h"
 #include "widgets/curves/curveeditor.h"
 #include "widgets/curves/curveeditorgroup.h"
