+++ rtgui/tools/icmpanel.cc
@@ -29,6 +29,7 @@
 #include "rtimage.h"
 #include "widgets/curves/curveeditor.h"
 #include "widgets/curves/curveeditorgroup.h"
+#include "widgets/basic/myexpander.h"
 
 #include "rtengine/dcp.h"
 #include "rtengine/iccstore.h"
