+++ rtgui/tools/locallabtools2.cc
@@ -29,6 +29,7 @@
 #include "rtengine/utils.h"
 #include "rtengine/rt_math.h"
 #include "labgrid.h"
+#include "widgets/basic/myexpander.h"
 
 #define MINNEIGH 0.1
 #define MAXNEIGH 1500
