+++ rtgui/tools/retinex.cc
@@ -6,6 +6,7 @@
 #include "eventmapper.h"
 #include "options.h"
 #include "rtimage.h"
+#include "widgets/basic/myexpander.h"
 #include "widgets/curves/curveeditor.h"
 #include "widgets/curves/curveeditorgroup.h"
 #include "widgets/curves/mycurve.h"
